requirements.txt
PySide6~=6.5.1.1
Pillow~=10.0.0